# Fix appendOrMergeRPC inefficiency in message size recalculation