## Summary

As discussed in #581 there an inefficiency in `appendOrMergeRPC` in calling `Size()` more times than needed.

## Fix

Instead of calling `lastRPC.Size()` that iterates over `RPC.Publish`, save the last known size and sum with a current message size and protobuf upper bound overhead.

## Status and Evaluation

Benchmark results
```
[master●●] % go test ./ -run ^$ -bench BenchmarkAppendOrMergeRPC -benchmem -count=10 > old.txt
[pavel/append-or-merge●] % go test ./ -run ^$ -bench BenchmarkAppendOrMergeRPC -benchmem -count=10 > new.txt
[pavel/append-or-merge●] % benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/libp2p/go-libp2p-pubsub
cpu: Apple M1 Pro
                          │   old.txt   │               new.txt               │
                          │   sec/op    │   sec/op     vs base                │
AppendOrMergeRPC/small-10   741.8n ± 0%   280.8n ± 0%  -62.15% (p=0.000 n=10)
AppendOrMergeRPC/large-10   36.88µ ± 0%   10.67µ ± 0%  -71.06% (p=0.000 n=10)
geomean                     5.231µ        1.731µ       -66.90%

                          │   old.txt    │                new.txt                │
                          │     B/op     │     B/op      vs base                 │
AppendOrMergeRPC/small-10     368.0 ± 0%     368.0 ± 0%       ~ (p=1.000 n=10) ¹
AppendOrMergeRPC/large-10   18.54Ki ± 0%   18.54Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                     2.581Ki        2.581Ki       +0.00%
¹ all samples are equal

                          │  old.txt   │               new.txt               │
                          │ allocs/op  │ allocs/op   vs base                 │
AppendOrMergeRPC/small-10   7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=10) ¹
AppendOrMergeRPC/large-10   216.0 ± 0%   216.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                     38.88        38.88       +0.00%
¹ all samples are equal
```