@MarcoPolo 2024-10-01
Generally looks good. Make sure to run the Fuzz tests as well. A benchmark might also be helpful here.