@algorandskiy 2025-05-20
The Fuzz looks working too well and finds broken input even on master branch (rerunning makes it pass tho):

```
go test -fuzz=FuzzAppendOrMergeRPC -run ^$ -v
=== RUN   FuzzAppendOrMergeRPC
fuzz: elapsed: 0s, gathering baseline coverage: 0/209 completed
fuzz: elapsed: 3s, gathering baseline coverage: 205/209 completed
fuzz: elapsed: 5s, gathering baseline coverage: 209/209 completed, now fuzzing with 10 workers
fuzz: elapsed: 6s, execs: 3044 (946/sec), new interesting: 0 (total: 209)
fuzz: elapsed: 9s, execs: 16393 (4450/sec), new interesting: 0 (total: 209)
fuzz: elapsed: 12s, execs: 24930 (2845/sec), new interesting: 0 (total: 209)
fuzz: elapsed: 15s, execs: 26633 (568/sec), new interesting: 0 (total: 209)
fuzz: elapsed: 17s, execs: 26897 (108/sec), new interesting: 0 (total: 209)
--- FAIL: FuzzAppendOrMergeRPC (17.46s)
    fuzzing process hung or terminated unexpectedly: exit status 2
    Failing input written to testdata/fuzz/FuzzAppendOrMergeRPC/fb45983a3dbe24d8
    To re-run:
    go test -run=FuzzAppendOrMergeRPC/fb45983a3dbe24d8
=== NAME
FAIL
exit status 1
FAIL	github.com/libp2p/go-libp2p-pubsub	17.895s
```

The same story is on my feature branch so I can't tell if it is a new or pre-existing issue.

I explained where "1+" comes from (it is a field key size from pb generator) and added a benchmark.