@algorandskiy 2025-05-21
@MarcoPolo could you retrigger the testing job, TestMessageBatchPublish timeout and it never happened on my local machines.