@algorandskiy 2025-05-22
Okay, it is 
```
=== RUN   TestMessageBatchPublish
2025/05/22 00:14:32 failed to sufficiently increase receive buffer size (was: 1024 kiB, wanted: 7168 kiB, got: 2048 kiB). See https://github.com/quic-go/quic-go/wiki/UDP-Buffer-Sizes for details.
panic: test timed out after 10m0s
	running tests:
		TestMessageBatchPublish (10m0s)
```