@algorandskiy 2025-05-22
> could I ask you to review it?

Absolutely!

> can you reach out via email so we can discuss details here?

Sent an email with details